+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 ## [0.0.5] 2018-08-27
 ### Bug fixing
 - Small changes for this warning: `Can't call setState (or forceUpdate) on an unmounted component.`
+- Updated README.md
+## Enhancement
 - Added support for collecting all states of the steps (Read the README.md for more)
+- Added support for scss (it auto-compiles into css)
+- Moved the styles from direct import (now you have to import the styles in your project)
+- Added new prop inside `steps` for adding an icon to the `stepName`
+- Small changes on how the components of the wizard are rendered
 
 ## [0.0.4] 2018-02-09
 ### Changes in tab margins